requirements.txt
# blocknet-mcp-server Dependencies

# Core MCP
mcp>=1.0.0

# Async HTTP
httpx>=0.27.0

# JSON
orjson>=3.10.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
structlog>=24.0.0

# Code Generation
pyyaml>=6.0.0
jinja2>=3.1.0

# Environment
python-dotenv>=1.0.0

# File watching
watchdog>=3.0.0
